requirements.txt
Flask>=2.0
python-dotenv
python-docx
mysql-connector-python
Werkzeug>=2.0
